major restructure, wip
