+++ pkg/internal/client/types.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package tfregistry
+package client
 
 import "time"
