+++ pkg/internal/tools/tools.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package tfregistry
+package tools
 
 import (
 	"net/http"
